// src/06-multi/code-a-la-mode/types.ts
export enum ItemPart {
  None = 'NONE',
  Dish = 'DISH',
  Blueberries = 'BLUEBERRIES',
  IceCream = 'ICE_CREAM',
  Strawberries = 'STRAWBERRIES',
  ChoppedStrawberries = 'CHOPPED_STRAWBERRIES',
  ChoppingBoard = 'CHOPPING_BOARD',
  Oven = 'OVEN',
  Dough = 'DOUGH',
  Croissant = 'CROISSANT',
  ChoppedDough = 'CHOPPED_DOUGH',
  RawTart = 'RAW_TART',
  Tart = 'TART',
  Window = 'WINDOW',
}

export enum TileType {
  Table = '#',
  Empty = '.',
  Dish = 'D',
  IceCream = 'I',
  Blueberries = 'B',
  Strawberries = 'S',
  ChoppingBoard = 'C',
  Oven = 'O',
  Dough = 'H',
  Window = 'W',
}